feature(SPR-639) add unit tests

+++ tests/conftest.py
@@ -2,20 +2,28 @@ import pytest
 from loguru import logger
 from sqlalchemy_utils.functions import drop_database
 
-from app import set_up_database_tables
+from app.database import Base, engine
+from app.dependencies import get_session_db
 from app.internal.config import DATABASE_URL
+
+
+@pytest.fixture(scope="session")
+def create_tables_db():
+    if "_test" in DATABASE_URL:
+        Base.metadata.create_all(bind=engine, checkfirst=True)
 
 
 @pytest.fixture(scope="session")
 def drop_db(request):
     @request.addfinalizer
     def drop_databases():
-        drop_database(DATABASE_URL)
+        if "_test" in DATABASE_URL:
+            drop_database(DATABASE_URL)
 
 
-@pytest.fixture(scope="session")
-def create_tables_db():
-    set_up_database_tables()
+@pytest.fixture
+def session_db():
+    return next(get_session_db())
 
 
 @pytest.fixture
